backend/requirements.txt
flask
flask-cors
azure-cosmos
PyJWT
bcrypt

